# specs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API enables end-users, owners and Sally devices to interact  with the central database.",
    "version": "1.0",
    "title": "Pantry REST API",
    "termsOfService": "http://casabots.com/terms/",
    "contact": {},
    "license": {
      "name": "Copyright (c) CASABOTS 2016"
    }
  },
  "host": "localhost:8080",
  "basePath": "/pantry",
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/add-custom-salads": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addCustomSalads",
        "operationId": "addCustomSaladsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "customSalads",
            "description": "customSalads",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CustomSalad"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CustomSalad"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/add-favorite-custom-salads": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addFavoriteCustomSalads",
        "operationId": "addFavoriteCustomSaladsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "salads",
            "in": "query",
            "description": "salads",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SaladList"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/add-favorite-salads": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addFavoriteSalads",
        "operationId": "addFavoriteSaladsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "salads",
            "in": "query",
            "description": "salads",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SaladList"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/custom-salads": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getCustomSalads",
        "operationId": "getCustomSaladsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CustomSalad"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/delete": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "deleteCurrentUser",
        "operationId": "deleteCurrentUserUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/favorite-salads": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getFavoriteSalads",
        "operationId": "getFavoriteSaladsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SaladList"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/fb-login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "fbLogin",
        "operationId": "fbLoginUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/get": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getCurrentUser",
        "operationId": "getCurrentUserUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/google-login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "googleLogin",
        "operationId": "googleLoginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idToken",
            "in": "query",
            "description": "idToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/remove-favorite-custom-salads": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "removeFavoriteCustomSalads",
        "operationId": "removeFavoriteCustomSaladsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "salads",
            "in": "query",
            "description": "salads",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SaladList"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/remove-favorite-salads": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "removeFavoriteSalads",
        "operationId": "removeFavoriteSaladsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "salads",
            "in": "query",
            "description": "salads",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SaladList"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/update": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/update-current": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "updateCurrentUser",
        "operationId": "updateCurrentUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getUser",
        "operationId": "getUserUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "deleteUser",
        "operationId": "deleteUserUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "customSalads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomSalad"
          }
        },
        "dob": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string"
        },
        "favouriteSalads": {
          "$ref": "#/definitions/SaladList"
        },
        "firstName": {
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/Character"
        },
        "lastName": {
          "type": "string"
        },
        "location": {
          "$ref": "#/definitions/GeoLocation"
        },
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "Character": {
      "type": "object"
    },
...
    "LoginUser": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "SaladList": {
      "type": "object",
      "properties": {
        "customSalads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomSalad"
          }
        },
        "preDefinedSalads": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "CustomSalad": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "GeoLocation": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "double"
        },
        "longitude": {
          "type": "number",
          "format": "double"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
  }
}