# src/main/java/com/casabots/pantry/model/CustomSalad.java
/*******************************************************************************
 * Copyright (c) CASABOTS 2016
 * All Rights Reserved
 ******************************************************************************/

package com.casabots.pantry.model;

import java.util.List;

public class CustomSalad {
    private String name;

    public CustomSalad() {
    }

    public CustomSalad(String name) {
        this.name = name;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

}